requirements.txt
anthropic>=0.50.0
streamlit>=1.40.0
python-dotenv>=1.0.0
